+++ CHANGELOG.md
@@ -25,6 +25,11 @@ The following changes have been completed, and will be included in the next tagg
 
 * Added `HashMap.forEach()`
 
+### Tests
+
+* `extractReasonFromCaught()` and `extractStackFromCaught()`
+  - unit tests no longer look for `Context.<anonymous>` in the stack trace
+
 ### Dependencies
 
 * Updated dev dependencies to resolve low-severity `npm audit` warnings.